app/src/main/java/com/tvmaze/tvseriessearch/data/source/network/TVMazeSourceProvider.kt
package com.tvmaze.tvseriessearch.data.source.network

import okhttp3.OkHttpClient
import retrofit2.Retrofit
import retrofit2.converter.gson.GsonConverterFactory

object TVMazeSourceProvider {

    private val client = OkHttpClient.Builder().build()

    private val retrofit = Retrofit.Builder()
        .baseUrl("https://api.tvmaze.com/")
        .addConverterFactory(GsonConverterFactory.create())
        .client(client)
        .build()

    fun<T> buildProvider(service: Class<T>): T {
        return retrofit.create(service);
    }
}